# expression: Optimize builtinArithmeticModRealSig and builtinGreatestDecimalSig using MergeNull method