<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary:

### What is changed and how it works?

What's Changed:

Use `MergeNulls()` in advance to avoid inefficient `SetNull()` operation as much as possible


### Related changes

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test

- Benchmark

```
go test --timeout 0 -v -benchmem -bench=BenchmarkVectorizedBuiltinCompareFunc -count=20 -run=BenchmarkVectorizedBuiltinCompareFunc | tee ~/old.txt
go test --timeout 0 -v -benchmem -bench=BenchmarkVectorizedBuiltinCompareFunc -count=20 -run=BenchmarkVectorizedBuiltinCompareFunc | tee ~/new.txt

go test --timeout 0 -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinCompareFunc | tee ~/old.txt
go test --timeout 0 -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinCompareFunc | tee ~/new.txt
benchstat old.txt new.txt
```

`builtinGreatestDecimalSig` bench result:
```
VectorizedBuiltinCompareFunc/builtinGreatestDecimalSig-VecBuiltinFunc-2         34.2µs ± 7%    32.2µs ± 4%    -5.89%  (p=0.000 n=20+19)
VectorizedBuiltinCompareFunc/builtinGreatestDecimalSig-NonVecBuiltinFunc-2      63.2µs ± 4%    63.1µs ± 4%      ~     (p=0.799 n=20+20
```

`builtinArithmeticModRealSig` bench result:
```
VectorizedBuiltinArithmeticFunc/builtinArithmeticModRealSig-VecBuiltinFunc#02-2                15.0µs ± 5%    13.8µs ± 4%   -8.38%  (p=0.000 n=20+19)
VectorizedBuiltinArithmeticFunc/builtinArithmeticModRealSig-NonVecBuiltinFunc#02-2             39.8µs ± 9%    38.6µs ± 8%   -3.02%  (p=0.014 n=20+20)
Vectoriz
```

### Release note <!-- bugfixes or new feature need a release note -->

- Optimize builtinArithmeticModRealSig and builtinGreatestDecimalSig using MergeNull method