@Tjianke 2021-01-18
/run-all-tests