@Tjianke 2021-02-22
/run-all-tests